requirements.txt
autopep8==1.4.4
mysql-connector==2.2.9
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
